refactor: socket 관련 type 별도 파일로 분리
