+++ server/src/socket/socket.d.ts
@@ -1,0 +1,8 @@
+declare module 'http' {
+  interface IncomingMessage {
+    session: session & {
+      nickName: string;
+    };
+    sessionID: string;
+  }
+}
